Modified same named entities to use unique table names.  Also corrected an NPE in test cleanup which occurred when embeddable metadata was present in the metadata repository.
git-svn-id: https://svn.apache.org/repos/asf/openjpa/trunk@828786 13f79535-47bb-0310-9956-ffa450edef68

+++ openjpa-persistence-jdbc/src/test/java/org/apache/openjpa/persistence/datacache/TestCacheExclusions.java
@@ -81,10 +81,12 @@ public class TestCacheExclusions extends AbstractCachedEMFTestCase {
             for (ClassMapping mapping : ((ClassMapping[]) emf
                 .getConfiguration().getMetaDataRepositoryInstance()
                 .getMetaDatas())) {
-                Query q =
-                    em.createNativeQuery("DROP TABLE "
-                        + mapping.getTable().getName());
-                q.executeUpdate();
+                if (mapping.getTable() != null) {
+                    Query q =
+                        em.createNativeQuery("DROP TABLE "
+                            + mapping.getTable().getName());
+                    q.executeUpdate();
+                }
             }
             em.getTransaction().commit();
             em.close();
